refactor: Dynamic IAM action list.[5]修正循环引用的问题
