+++ src/ac/iam/iam.go
@@ -414,6 +414,6 @@ func (a *authorizer) BatchRegisterResourceCreatorAction(ctx context.Context, h h
 	return a.authClientSet.BatchRegisterResourceCreatorAction(ctx, h, input)
 }
 
-func (a *authorizer) CreateModelInstanceActions(ctx context.Context, h http.Header, input []ResourceAction) error {
+func (a *authorizer) CreateModelInstanceActions(ctx context.Context, h http.Header, input []metadata.Object) error {
 	return a.authClientSet.CreateModelInstanceActions(ctx, h, input)
 }
