+++ src/apimachinery/authserver/authserver.go
@@ -17,7 +17,6 @@ import (
 	"fmt"
 	"net/http"
 
-	"configcenter/src/ac/iam"
 	"configcenter/src/ac/meta"
 	"configcenter/src/apimachinery/rest"
 	"configcenter/src/apimachinery/util"
@@ -35,7 +34,7 @@ type AuthServerClientInterface interface {
 		[]metadata.IamCreatorActionPolicy, error)
 	BatchRegisterResourceCreatorAction(ctx context.Context, h http.Header, input metadata.IamInstancesWithCreator) (
 		[]metadata.IamCreatorActionPolicy, error)
-	CreateModelInstanceActions(ctx context.Context, h http.Header, input []iam.ResourceAction) error
+	CreateModelInstanceActions(ctx context.Context, h http.Header, input []metadata.Object) error
 }
 
 func NewAuthServerClientInterface(c *util.Capability, version string) AuthServerClientInterface {
